restructure cache chunk generation, add LFU policy delete
